+++ js/main.js
@@ -17,7 +17,7 @@ import { setRetroTab, retroPrevQ, retroNextQ } from './retrospectiva.js';
 import { saveFinMes } from './financeiro.js';
 import { shareAchievement } from './share.js';
 import { showLegal, closeLegal, setLegalTab } from './legal.js';
-import { startRotinaWizard, closeRotinaWizard, rwCheckStep1, rwNext, rwBack, rwAddMateria, rwRemoveMateria, rwAddChecklist, rwRemoveChecklist, rwToggleDia, rwSetDuracao, rwGenerate, rwRemoveBloco, rwConfirm, toggleRotinaBloco, toggleRotinaChecklist, encerrarRotina } from './rotina.js';
+import { startRotinaWizard, startRotinaWizardEdit, closeRotinaWizard, rwCheckStep1, rwNext, rwBack, rwAddMateria, rwRemoveMateria, rwAddChecklist, rwRemoveChecklist, rwToggleDia, rwSetDuracao, rwGenerate, rwRemoveBloco, rwConfirm, toggleRotinaBloco, toggleRotinaChecklist, encerrarRotina } from './rotina.js';
 
 // Expõe ao window para os onclick inline no HTML
 window.clearFieldErr = clearFieldErr;
@@ -90,6 +90,7 @@ window.closeLegal = closeLegal;
 window.setLegalTab = setLegalTab;
 
 window.startRotinaWizard = startRotinaWizard;
+window.startRotinaWizardEdit = startRotinaWizardEdit;
 window.closeRotinaWizard = closeRotinaWizard;
 window.rwCheckStep1 = rwCheckStep1;
 window.rwNext = rwNext;